src/requirements.txt
python-gnupg==0.4.6
PyYAML==5.3.1
boto3==1.14.1
botocore==1.17.1
